requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
google-generativeai==0.8.4
python-dotenv==1.0.1
gunicorn==21.2.0
jinja2==3.1.3
aiosqlite==0.20.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
